NFDIV-3751: move divorce payment domain model objects
